Renderizado del mapa en web ahora posible
